# db/postgres/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.22.0

package dbgen

import (
	"database/sql"
)

type Admin struct {
	ID             int64
	FirstName      sql.NullString
	LastName       sql.NullString
	Username       string
	Email          sql.NullString
	PasswordHashed sql.NullString
	PasswordSalt   sql.NullString
	Status         sql.NullString
	CreatedBy      int64
	UpdatedBy      int64
	DeletedBy      int64
	CreatedAt      sql.NullTime
	UpdatedAt      sql.NullTime
	DeletedAt      sql.NullTime
}

type Category struct {
	ID          int64
	Description sql.NullString
	Title       sql.NullString
	CreatedBy   int64
	UpdatedBy   int64
	DeletedBy   int64
	CreatedAt   sql.NullTime
	UpdatedAt   sql.NullTime
	DeletedAt   sql.NullTime
}

type Device struct {
	ID          int64
	DeviceToken sql.NullString
	Platform    sql.NullString
	Status      sql.NullString
	CreatedBy   int64
	UpdatedBy   int64
	DeletedBy   int64
	CreatedAt   sql.NullTime
	UpdatedAt   sql.NullTime
	DeletedAt   sql.NullTime
}

type Object struct {
	ID         int64
	CategoryID int64
	Content    sql.NullString
	Title      sql.NullString
	CreatedBy  int64
	UpdatedBy  int64
	DeletedBy  int64
	CreatedAt  sql.NullTime
	UpdatedAt  sql.NullTime
	DeletedAt  sql.NullTime
}

type SchemaMigration struct {
	Version string
}

type User struct {
	ID             int64
	FirstName      sql.NullString
	LastName       sql.NullString
	Username       string
	Email          sql.NullString
	PasswordHashed sql.NullString
	PasswordSalt   sql.NullString
	Status         sql.NullString
	CreatedBy      int64
	UpdatedBy      int64
	DeletedBy      int64
	CreatedAt      sql.NullTime
	UpdatedAt      sql.NullTime
	DeletedAt      sql.NullTime
}